OscoNet

Documentation for OscoNet.

OscoNet.bootstrap_hypothesis_test

OscoNet.cooscillation_distance

OscoNet.create_edge_network

OscoNet.find_best_psi_for_each_gene_pair

OscoNet.find_minimum_distance

OscoNet.get_metrics_for_different_qvalue_thresholds

OscoNet.get_permuted_cost

OscoNet.get_pvalues

OscoNet.qvalue_estimate

OscoNet.simulate_data

OscoNet.true_adj_matrix

OscoNet.vec_triu_loop

OscoNet.bootstrap_hypothesis_test — Method

Arguments

Returns

OscoNet.cooscillation_distance — Method

Distance function for two co-oscillating genes across N cells. We assume genes have identical sinusoidal expression profiles, except for a phase shift Ψ.

Arguments

X: Expression of gene X from N cells

Y: Expression of gene Y from the same N cells

cosΨ::AbstractFloat: Cosine of the phase shift, Ψ

Returns

d::AbstractFloat: Distance between the two genes for a given Ψ

OscoNet.create_edge_network — Method

CreateEdgeNetwork - Create Edge file. This is needed before hypothesis test q-value derived adjacency matrix can be consumed by R network analysis code. Return a pandas dataframe with 3 columns, two gene names for the gene-pair and the cost value

OscoNet.find_best_psi_for_each_gene_pair — Method

Arguments

Returns

OscoNet.find_minimum_distance — Method

Find the minimum distance between two genes X and Y, and the value of Ψ that minimises it.

Arguments

X: Expression of gene X from N cells

Y: Expression of gene Y from the same N cells

Returns

minimum::AbstractFloat: Minimum distance between gene X and Y

minimiser::AbstractFloat: Value of Ψ that minimises the distance

OscoNet.get_metrics_for_different_qvalue_thresholds — Method

Get true positive, false discovery and false positive rates :param qvalues: qvalue G X G matrix where G=number of genes :param adjMatrixtrue: GXG binary adjacency matrix of true gene pairs :param alphavalues: one dimensional vector for threshold values :return: TPR, FDR, FPR vector of same size as alpha_values

OscoNet.get_permuted_cost — Method

Arguments

data::: G X N tensor of gene expression

n_permutations::Int: number of bootstrap permutations

Returns

cost_permuted::Array{Float64}: permuted cost

OscoNet.get_pvalues — Method

Get pvalues from bootstrap

Arguments

cost_unpermuted: G X G - only upper triangular used

cost_permuted: G X G X n permutations

#Returns

pvalues

OscoNet.qvalue_estimate — Method

Estimates qvalues from pvalues

Arguments

pvalues:

m: number of tests. If not specified m = length(pvalues)

verbose::Bool: print verbose messages (default is false)

pi0: if None, it's estimated as suggested in Storey and Tibshirani, 2003.

OscoNet.simulate_data — Method

Generate synthetic data. Reproduced from the original Oscope paper

Included below is the original description from the supplementary material of that paper:

Sim I: Oscope paper supplementary

1000 genes and 100 cells. 90 out of the 1000 genes were simulated as oscillators. The 90 oscillators were simulated in 3 frequency groups, each group contains 30 genes. Group 1 and 3 follow the same order, while genes in group 2 follow another order.

The relative speeds of the 3 groups are proportional to 2:3:6. Within each frequency group, genes were further simulated with strong and weak signals. Half of the oscillatory genes were simulated as strong oscillators with σ_g = σ_str. The other half were simulated as weak oscillators with σ_g = 2σ_str. Starting Ψ Ψ_g varies in different genes within a frequency group. The remaining genes except the oscillators are called noise genes. Noise genes were simulated as random Gaussian noise. The noise level was adjusted to be comparable to the average noise signal among all oscillators. The σ_str varies from 0.05 to 0.4 in 5 steps.

Arguments

half_group::Int=15: Half the size of each gene group. For example if half_group=3, each group will have 6 co-oscillating genes, 3 of which

will be strong oscillators and 3 weak (double the noise)

n_genes::Int=1000: Total number of genes

n_cells::Int=100: Total number of cells

noise_level::Int=1: Noise level index (1 to 6)

n_groups::Int=3: Number of groups

Returns

data::Matrix{<:AbstractFloat}: Simulated gene expression data with shape n_genes x n_cells

Ψ_g::Vector{<:AbstractFloat: Vector of phases for each gene

ω::Vector{<:AbstractFloat}: Vector of angular speeds for each gene

OscoNet.true_adj_matrix — Method

Construct the ground truth adjacency matrix for a given numer of genes and ω vector. Intended for use with output from the simulate_data function.

Arguments

n_genes::Integer: Total number of genes

ω::Vector{<:AbstractFloat}: Vector of angular speeds

Returns

adj_mat::Matrix{Bool}: Boolean matrix which is 1 if genes co-oscillate, and 0 otherwise

OscoNet.vec_triu_loop — Method

A function to return the elements of the (strict) upper triangle of a matrix, M, as a vector