Internals

Documentation for OscoNet.jl's internal functions.

See the Public Documentation section for documentation of the public interface.

Index

OscoNet.find_best_psi_for_each_gene_pair

OscoNet.get_metrics_for_different_qvalue_thresholds

OscoNet.get_permuted_cost

OscoNet.get_pvalues

OscoNet.qvalue_estimate

OscoNet.vec_triu_loop

Internal Interface

OscoNet.find_best_psi_for_each_gene_pair — Method

Arguments

Returns

OscoNet.get_metrics_for_different_qvalue_thresholds — Method

Get true positive, false discovery and false positive rates :param qvalues: qvalue G X G matrix where G=number of genes :param adjMatrixtrue: GXG binary adjacency matrix of true gene pairs :param alphavalues: one dimensional vector for threshold values :return: TPR, FDR, FPR vector of same size as alpha_values

OscoNet.get_permuted_cost — Method

Arguments

data::: G X N tensor of gene expression

n_permutations::Int: number of bootstrap permutations

Returns

cost_permuted::Array{Float64}: permuted cost

OscoNet.get_pvalues — Method

Get pvalues from bootstrap

Arguments

cost_unpermuted: G X G - only upper triangular used

cost_permuted: G X G X n permutations

#Returns

pvalues

OscoNet.qvalue_estimate — Method

Estimates qvalues from pvalues

Arguments

pvalues:

m: number of tests. If not specified m = length(pvalues)

verbose::Bool: print verbose messages (default is false)

pi0: if None, it's estimated as suggested in Storey and Tibshirani, 2003.

OscoNet.vec_triu_loop — Method

A function to return the elements of the (strict) upper triangle of a matrix, M, as a vector