Internals

Documentation for OscoNet.jl's internal functions.

See the Public Documentation section for documentation of the public interface.

Index

OscoNet.construct_symmetric_matrix_from_upper_triangular_vector

OscoNet.find_best_psi_for_each_gene_pair

OscoNet.get_metrics_for_different_qvalue_thresholds

OscoNet.get_permuted_cost

OscoNet.get_pvalues

OscoNet.istriangular

OscoNet.qvalue_estimate

OscoNet.vec_triu_loop

Internal Interface

OscoNet.construct_symmetric_matrix_from_upper_triangular_vector — Method

Construct a symmetric matrix with Inf diagonal values using a vector whose length is a triangular number. This vector is intended to be the output of vec_triu_loop.

Examples

julia> using OscoNet

julia> A = [1 2 3; 4 5 6; 7 8 9]
3×3 Matrix{Int64}:
 1  2  3
 4  5  6
 7  8  9

julia> A_flatten = OscoNet.vec_triu_loop(A)
3-element Vector{Int64}:
 2
 3
 6

julia> OscoNet.construct_symmetric_matrix_from_upper_triangular_vector(A_flatten)
3×3 Matrix{Float64}:
 Inf    2.0   3.0
  2.0  Inf    6.0
  3.0   6.0  Inf

OscoNet.find_best_psi_for_each_gene_pair — Method

Compute the minimum distance between gene x and gene y, for a set of genes x, y ∈ X.

Arguments

data::Matrix: Matrix of size (ngenes x ncells), whose elements represent gene expression.

Returns

Ψ::Array{Float64}: Array of size (ngenes x ngenes), whose (x,y)-th element

represents the phase shift Ψ between gene x and gene y.

distance::Array{Float64}: Array of size (ngenes x ngenes), whose (x,y)-th element

represents the minimum distance d between gene x and gene y.

OscoNet.get_metrics_for_different_qvalue_thresholds — Method

Get true positive, false discovery and false positive rates :param qvalues: qvalue G X G matrix where G=number of genes :param adjMatrixtrue: GXG binary adjacency matrix of true gene pairs :param alphavalues: one dimensional vector for threshold values :return: TPR, FDR, FPR vector of same size as alpha_values

OscoNet.get_permuted_cost — Method

Compute the minimum distance between gene x and the n-th random permutation of gene y, for a set of genes x, y ∈ X.

Arguments

data::Matrix: Matrix of size (ngenes x ncells), whose elements represent gene expression.

n_permutations::Int: number of bootstrap permutations

Returns

cost_permuted::Array{Float64}: Array of size (ngenes x ngenes x n_permutations), whose (x,y,n)-th element

represents the minimum distance between gene x and the n-th random permutation of gene y.

OscoNet.get_pvalues — Method

Get pvalues from bootstrap

Arguments

cost_unpermuted: G X G - only upper triangular used

cost_permuted: G X G X n permutations

#Returns

pvalues

OscoNet.istriangular — Method

istriangular(x::Number) -> Bool

Test whether x is a triangular number.

By solving the quadratic formula for the the m-th triangular number n = m(m-1)/2, it can be shown that an integer is triangular if and only if 8n+1 is a perfect square. In other words, if and only if sqrt(8n+1) is an integer.

Examples

julia> using OscoNet

julia> OscoNet.istriangular(45.0)
true

OscoNet.qvalue_estimate — Method

Estimates qvalues from pvalues

Arguments

pvalues:

m: number of tests. If not specified m = length(pvalues)

verbose::Bool: print verbose messages (default is false)

pi0: if None, it's estimated as suggested in Storey and Tibshirani, 2003.

OscoNet.vec_triu_loop — Method

A function to return the elements of the (strict) upper triangle of a matrix, M, as a vector.

Examples

julia> using OscoNet

julia> A = [1 2 3; 4 5 6; 7 8 9]
3×3 Matrix{Int64}:
 1  2  3
 4  5  6
 7  8  9

julia> OscoNet.vec_triu_loop(A)
3-element Vector{Int64}:
 2
 3
 6